Step three: restructure and untangle Zone IPCs
