+++ src/ipc/zone/server/currency_info.rs
