+++ src/ipc/zone/server/effect_result.rs
